backend/requirements.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-decouple==3.8
markdown==3.5.2
django-filter==23.5

gunicorn==21.2.0
dj-database-url==2.1.0
whitenoise==6.6.0
